Update modifiedBy field with userDetails.userName in Managerequest.vue and ManagerequestSafety.vue; add userDetails to data in Myrequest2.vue
The code diff above shows changes made in three Vue components: Managerequest.vue, ManagerequestSafety.vue, and Myrequest2.vue.

In Managerequest.vue and ManagerequestSafety.vue, the "modifiedBy" field in various API requests has been updated to use "this.userDetails.userName" instead of the original value. This change was made in multiple places throughout each file.

In Myrequest2.vue, "this.userDetails" has been added to the data function, but it is not clear from the diff whether it is used elsewhere in the component.

Overall, the changes made in this diff are relatively minor and seem to be related to updating the "modifiedBy" field with the current user's name.